videos test file works for basic jw player videos
no thumbnails shown yet but one is correctly referenced, it did work at one time on another site. This is the first attempt at duplicating it on widgetmash.com , there are thumbnails .jpgs missing and other directory trees which are not on my computer for some reason, will be looking around for it. stay tuned. th